+++ seaborn/seaborn_table.py
@@ -1,3 +1,0 @@
-# this is for backwards compatibility
-from seaborn_table.table import SeabornTable, SeabornRow
-